+++ src/support/slack/actions/index.js
@@ -25,6 +25,8 @@ import {
 import { onboardSiteModal, startOnboarding } from './onboard-modal.js';
 import { preflightConfigModal } from './preflight-config-modal.js';
 import openPreflightConfig from './open-preflight-config.js';
+import { openSetImsOrgModal, setImsOrgModal } from './set-ims-org-modal.js';
+import { openAddSiteModal, addSiteModal } from './add-site-modal.js';
 
 const actions = {
   approveFriendsFamily,
@@ -35,10 +37,14 @@ const actions = {
   onboardSiteModal,
   onboardLLMOModal,
   updateIMSOrgModal,
+  setImsOrgModal,
+  addSiteModal,
   start_onboarding: startOnboarding,
   start_llmo_onboarding: startLLMOOnboarding,
   preflight_config_modal: preflightConfigModal,
   open_preflight_config: openPreflightConfig,
+  open_set_ims_org_modal: openSetImsOrgModal,
+  open_add_site_modal: openAddSiteModal,
   add_entitlements_action: addEntitlementsAction,
   update_org_action: updateOrgAction,
 };