+++ src/support/slack/actions/onboard-modal.js
@@ -210,6 +210,7 @@ export function startOnboarding(lambdaContext) {
                     { text: 'Paid', value: 'paid' },
                     { text: 'PLG', value: 'plg' },
                     { text: 'Test', value: 'test' },
+                    { text: 'Dummy', value: 'dummy' },
                   ];
 
                   // Use provided profile or default to 'demo'
@@ -265,6 +266,13 @@ export function startOnboarding(lambdaContext) {
                       text: 'Test',
                     },
                     value: 'test',
+                  },
+                  {
+                    text: {
+                      type: 'plain_text',
+                      text: 'Dummy',
+                    },
+                    value: 'dummy',
                   },
                 ],
               },